refactor: upgrade constants
