+++ src/main/java/ru/yandex/practicum/filmorate/constans/HomeDir.java
@@ -1,0 +1,8 @@
+package ru.yandex.practicum.filmorate.constans;
+
+public class HomeDir {
+    private HomeDir() {
+    }
+
+    public static final String HOME = System.getProperty("user.dir") + "/src/main/resources/";
+}
