+++ src/main/java/ru/yandex/practicum/filmorate/constans/NameLogs.java
@@ -1,0 +1,11 @@
+package ru.yandex.practicum.filmorate.constans;
+
+public class NameLogs {
+    private NameLogs() {
+    }
+
+    public static final String SERVICE_IN_DB = "Сервис выполняет запрос в БД";
+    public static final String SERVICE_FROM_DB = "В сервис из БД вернулся";
+    public static final String CLIENT_SEND_REQUEST = "Клиент сделал запрос";
+    public static final String DB_RUNNING = "В БД выполняется запрос";
+}
