+++ src/main/java/ru/yandex/practicum/filmorate/utils/BanListUserName.java
@@ -3,12 +3,6 @@
  *
  * DangerZone!!!
  *
- * SET_MIN_DATE
- * Общий запрет внесения фильма, раньше определенной даты.
- * При изменении настройки SET_MIN_DATE, необходимо изменить сообщение
- * о событии ReleaseDateFilms. Сделать это можно в файле с названием:
- * ValidationMessages.properties. Ключ - ReleaseDateFilms.invalid
- *
  * BAN_LIST_ADD_LOGIN
  * Список логинов, которые запрещены для регистрации и обновления.
  * Эндпоинты: createUser("/users") и updateUser("/users")
@@ -21,27 +15,20 @@
  *
  */
 
-package ru.yandex.practicum.filmorate.constans;
+package ru.yandex.practicum.filmorate.utils;
 
 import java.io.IOException;
-import java.time.LocalDate;
-import java.time.Month;
 import java.util.List;
 import java.util.stream.Collectors;
 
 import static java.nio.file.Files.lines;
 import static java.nio.file.Paths.get;
+import static ru.yandex.practicum.filmorate.constans.HomeDir.HOME;
 
-public class Settings {
-    private Settings() {
+public class BanListUserName {
+    private BanListUserName() {
     }
 
-    public static final String SERVICE_IN_DB = "Сервис выполняет запрос в БД";
-    public static final String SERVICE_FROM_DB = "В сервис из БД вернулся";
-    public static final String CLIENT_SEND_REQUEST = "Клиент сделал запрос";
-    public static final String DB_RUNNING = "В БД выполняется запрос";
-    public static LocalDate SET_MIN_DATE = LocalDate.of(1895, Month.DECEMBER, 28);
-    private static final String HOME = System.getProperty("user.dir") + "/src/main/resources/";
     public static List<String> BAN_LIST_ADD_LOGIN;
     public static List<String> BAN_LIST_FIND_LOGIN;
 
@@ -53,4 +40,5 @@ public class Settings {
             throw new RuntimeException(e);
         }
     }
+
 }
